+++ src/test/java/com/ditrit/letomodelizerapi/cucumber/StepDefinitions.java
@@ -441,6 +441,11 @@ public class StepDefinitions {
         this.clean("ai/secrets", String.format("key=%s", key));
     }
 
+    @And("I clean AI configuration {string}")
+    public void cleanConfiguration(String key) throws URISyntaxException, IOException, InterruptedException  {
+        this.clean("ai/configurations", String.format("key=%s", key));
+    }
+
     public void clean(String entity, String query) throws URISyntaxException, IOException, InterruptedException {
         this.request(String.format("/%s?%s", entity, query));
         if (statusCode == 200 && json.get("totalElements").asInt() > 0) {